more self review
