+++ packages/plugin-docs-cli/src/validation/rules/utils.ts
@@ -1,0 +1,21 @@
+/**
+ * Returns a set of 1-based line numbers inside fenced code blocks.
+ */
+export function getCodeBlockLines(content: string): Set<number> {
+  const lines = content.split('\n');
+  const codeLines = new Set<number>();
+  let inCodeBlock = false;
+
+  for (let i = 0; i < lines.length; i++) {
+    if (/^```/.test(lines[i].trim())) {
+      inCodeBlock = !inCodeBlock;
+      codeLines.add(i + 1);
+      continue;
+    }
+    if (inCodeBlock) {
+      codeLines.add(i + 1);
+    }
+  }
+
+  return codeLines;
+}
